Rename UserDetails to Authentication
